pyproject.toml
[project]
name = "ana-website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "flask-sqlalchemy>=3.1.1",
    "pydantic>=2.12.5",
    "pyjwt>=2.12.1",
    "requests>=2.33.1",
]
